% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium.R
\name{gen_selen}
\alias{gen_selen}
\title{Generate an instance of self-contained Selenium instance}
\usage{
gen_selen(headless = TRUE)
}
\arguments{
\item{headless}{whether to generate a headless instance (no GUI)}
}
\value{
an instance of Selenium instance, a list containing both the Selenium Server and the Remote Driver
}
\description{
Generate an instance of self-contained Selenium instance
}
