% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{generate_hash}
\alias{generate_hash}
\title{Genrate a random hash}
\usage{
generate_hash(ending = ".RDS", length = 40)
}
\arguments{
\item{ending}{the ending of the hash}

\item{length}{the length of the random part}
}
\value{
a hash
}
\description{
This function generates a random hash that can be used as a burner file name
}
\author{
Chung-hong Chan
}
