% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{get_links}
\alias{get_links}
\title{Get links (urls) from the DB}
\usage{
get_links(
  fname = NULL,
  size = 300,
  safe = FALSE,
  collection = "articles",
  db = "main",
  unsafe_pubs = c("Zeit", "Saarbrücker Zeitung", "RT deutsch")
)
}
\arguments{
\item{fname}{file name of the RDS file. If null, the function returns a vector of links}

\item{size}{number of links to fetch}

\item{safe}{whether to only get links from "safe" sources}

\item{collection}{Mongodb collection}

\item{db}{MongoDB db}

\item{unsafe_pubs}{a vector of publications to avoid}
}
\value{
either a vector of links (if \code{fname} is null) or the \code{fname} invisibly
}
\description{
This function gets links (urls) from the DB where the field \code{htmlfile} is empty, i.e. not yet scraped.
}
\author{
Chung-hong Chan
}
