% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_docdb.R
\name{insert_doc_link}
\alias{insert_doc_link}
\title{Insert links to MongoDB and return new links
This function inserts links (from either the RSS aggregator or archive scrappers) into the MongoDB (default to main.articles).}
\usage{
insert_doc_link(links, debug = FALSE, db = "main", collection = "articles")
}
\arguments{
\item{links, }{a data.frame / tibble of the docdb format: pub, link, pubdate, title, description}

\item{debug}{logical, whether debug message should be displayed}

\item{db}{MongoDB db}

\item{collection}{Mongodb collection}
}
\value{
A docdb-formated tibble of new links, which are not yet in the collection before the insertion
}
\description{
Insert links to MongoDB and return new links
This function inserts links (from either the RSS aggregator or archive scrappers) into the MongoDB (default to main.articles).
}
\author{
Chung-hong Chan
}
