% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{modify_url}
\alias{modify_url}
\title{Modify urls}
\usage{
modify_url(url, rm_query = TRUE, replace_domains = NULL)
}
\arguments{
\item{url}{character, input url}

\item{rm_query}{logical, whether to remove the queries (?) and fragments (#)}

\item{replace_domains}{character vector, when is not NULL, one domain from this vector is chosen to replace the hostname of the input url}
}
\value{
a modified url
}
\description{
Modify urls, either by removing the query part, or the domain names.
}
\author{
Chung-hong Chan
}
