% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{pack_work}
\alias{pack_work}
\title{Pack the job from \code{work_dl}}
\usage{
pack_work(output_file = "job.tar.gz", rds, output_dir, delete = FALSE)
}
\arguments{
\item{output_file}{archive}

\item{rds}{rds file}

\item{output_dir}{html directory}

\item{delete}{whether to delete \code{rds} and \code{output_dir} after packing}
}
\value{
\code{output_file}, invisibily.
}
\description{
Pack the job from \code{work_dl} as a single tar.gz
}
