% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_docdb.R
\name{push_html}
\alias{push_html}
\title{Insert a single url and html file to MongoDB
This function takes the output from \code{\link[=scrape]{scrape()}} and inserts it to the MongoDB (default to main.articles and main.html). Concretely, it updates the \code{htmlfile} field of the matching records and pushes the HTML files in \code{output_dir} into MongoDB's GridFS.}
\usage{
push_html(
  url,
  fname,
  output_dir = Sys.getenv("ARTICLE_DIR"),
  db = "main",
  collection = "articles",
  prefix = "S3://fodira-html/",
  delete = TRUE
)
}
\arguments{
\item{url}{a url}

\item{fname}{file name of the HTML file}

\item{output_dir}{a directory to hold HTML files}

\item{db}{MongoDB db}

\item{collection}{MongoDB collection}

\item{prefix}{S3 path}

\item{delete}{whether to delete the HTML files afterwards.}
}
\value{
The file name uploaded to GridFS, return NA if \code{fname} doesn't exist
}
\description{
Insert a single url and html file to MongoDB
This function takes the output from \code{\link[=scrape]{scrape()}} and inserts it to the MongoDB (default to main.articles and main.html). Concretely, it updates the \code{htmlfile} field of the matching records and pushes the HTML files in \code{output_dir} into MongoDB's GridFS.
}
\author{
Chung-hong Chan
}
