% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_docdb.R
\name{push_html_job}
\alias{push_html_job}
\title{Insert output from \code{\link[=pack_work]{pack_work()}} to MongoDB
This function takes the output from \code{\link[=pack_work]{pack_work()}} and inserts it to the MongoDB (default to main.articles and main.html). Concretely, it updates the \code{htmlfile} field of the matching records and pushes the HTML files into MongoDB's GridFS.}
\usage{
push_html_job(
  job_fname,
  db = "main",
  collection = "articles",
  prefix = "html",
  delete = TRUE
)
}
\arguments{
\item{job_fname}{path of the job (a tar.gz file) from \code{\link[=pack_work]{pack_work()}}}

\item{db}{MongoDB db}

\item{collection}{MongoDB collection}

\item{prefix}{MongoDB GridFS prefix}

\item{delete}{whether to delete the HTML files afterwards.}
}
\description{
Insert output from \code{\link[=pack_work]{pack_work()}} to MongoDB
This function takes the output from \code{\link[=pack_work]{pack_work()}} and inserts it to the MongoDB (default to main.articles and main.html). Concretely, it updates the \code{htmlfile} field of the matching records and pushes the HTML files into MongoDB's GridFS.
}
\author{
Chung-hong Chan
}
