% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{request_links}
\alias{request_links}
\title{Request some links from host}
\usage{
request_links(
  fname = generate_hash(),
  host = Sys.getenv("FODIRA_HOST"),
  size = 300,
  safe = FALSE,
  verbose = TRUE,
  check = TRUE,
  exact_fname
)
}
\arguments{
\item{fname}{file name of the job, default to a random hash}

\item{host}{host string in the format of "username@ipaddress", default to the envvar "FODIRA_HOST"}

\item{size}{number of links to fetch}

\item{safe}{whether to only get links from "safe" sources}

\item{verbose}{whether to display messages}

\item{check}{whether to check for the existence of the file \code{fname} after the SSH session}

\item{exact_fname}{exact file to be collect from host}
}
\value{
the actual path of the saved RDS file
}
\description{
This is a function mostly for workers to request links from the remote host with the DB. It runs \code{\link[=get_links]{get_links()}} remotely and transfer the RDS file locally to the current \code{\link[=tempdir]{tempdir()}}. Please make sure you have passwordless access to the remote host, i.e. your public key is in the remote host.
}
\author{
Chung-hong Chan
}
