% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{submit_job}
\alias{submit_job}
\title{Submit the output from \code{\link[=pack_work]{pack_work()}} back to host}
\usage{
submit_job(
  job_fname,
  host = Sys.getenv("FODIRA_HOST"),
  verbose = FALSE,
  check = TRUE,
  delete = TRUE,
  host_directory = "finished_jobs"
)
}
\arguments{
\item{job_fname}{job file name from \code{\link[=pack_work]{pack_work()}}}

\item{host}{host string in the format of "username@ipaddress", default to the envvar "FODIRA_HOST"}

\item{verbose}{whether to display messages}

\item{check}{whether to check for the existence of the file \code{fname} after the SSH session}

\item{delete}{whether to delete \code{job_fname} after submission}

\item{host_directory}{path to upload to in the host}
}
\value{
return 0 if everything is alright
}
\description{
This is a function mostly for workers to return a job packed with \code{\link[=pack_work]{pack_work()}} back to the host. Please make sure you have passwordless access to the remote host, i.e. your public key is in the remote host. By default, it will return to the "finished_job" directory in the host
}
\author{
Chung-hong Chan
}
