% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{harmonize_output}
\alias{harmonize_output}
\title{Harmonize output from either rss aggregator or archive scrapers}
\usage{
harmonize_output(
  output,
  pub = c("nordbayern.de", "Spiegel", "Bild", "Berliner Zeitung",
    "Stuttgarter Zeitung", "HNA", "Merkur", "Stern", "Klasse gegen Klasse",
    "Tagesspiegel", "DieUnbestechlichen", "Heute"),
  remove_duplicates = TRUE
)
}
\arguments{
\item{output}{data.frame in the standardized docdb format}

\item{pub}{a vector of outlets requiring cleaning}

\item{remove_duplicates}{whether to remove deplicated entries with duplicated links}
}
\value{
a cleaned data.frame in the docdb format
}
\description{
This harmonizes the output by cleaning the links of some selected outlets and remove entries with duplicated links. The cleaning involves the removal of query strings in links using \code{\link[=modify_url]{modify_url()}}.
}
