% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_docdb.R
\name{push_html}
\alias{push_html}
\title{Insert output and html files from scrape() to MongoDB
This function takes the output from \code{\link[=scrape]{scrape()}} and inserts it to the MongoDB (default to main.articles and main.html). Concretely, it updates the \code{htmlfile} field of the matching records and pushes the HTML files in \code{output_dir} into MongoDB's GridFS.}
\usage{
push_html(
  output,
  output_dir = Sys.getenv("ARTICLE_DIR"),
  db = "main",
  collection = "articles",
  prefix = "html",
  delete = TRUE
)
}
\arguments{
\item{output}{data.frame from \code{\link[=scrape]{scrape()}}}

\item{output_dir}{a directory to hold HTML files}

\item{db}{MongoDB db}

\item{collection}{MongoDB collection}

\item{prefix}{MongoDB GridFS prefix}

\item{delete}{whether to delete the HTML files afterwards.}
}
\value{
a vector of file names uploaded to GridFS (invisibly)
}
\description{
Insert output and html files from scrape() to MongoDB
This function takes the output from \code{\link[=scrape]{scrape()}} and inserts it to the MongoDB (default to main.articles and main.html). Concretely, it updates the \code{htmlfile} field of the matching records and pushes the HTML files in \code{output_dir} into MongoDB's GridFS.
}
\author{
Chung-hong Chan
}
